ngo-crawling/requirements.txt
Flask==2.0.2
requests==2.26.0
requests-toolbelt==0.9.1
beautifulsoup4==4.10.0
python-dotenv==0.19.1